+++ zcash_client_sqlite/src/wallet/scanning.rs
@@ -15,6 +15,7 @@ use zcash_protocol::{
     ShieldedProtocol,
 };
 
+use crate::wallet::refresh_status_requests;
 use crate::TableConstants;
 use crate::{
     error::SqliteClientError,
@@ -548,12 +549,16 @@ pub(crate) fn update_chain_tip<P: consensus::Parameters>(
         None => tip_entry.block_range().clone(),
     };
 
+    // persist the updated scan queue entries
     replace_queue_entries::<SqliteClientError>(
         conn,
         &query_range,
         tip_shard_entry.into_iter().chain(Some(tip_entry)),
         false,
     )?;
+
+    // ensure that transaction status requests exist for any unmined, unexpired transactions
+    refresh_status_requests(conn)?;
 
     Ok(())
 }